Moved conditions to types.lisp, added ssh-unknown-hostkey
The new condition ssh-unknown-hostkey is signalled, when the host key
of the server to which we are connecting is not in the (possibly newly
created and thus empty) known-hosts file. The restart accept-always
writes that key to the known-hosts file and accept-once just accepts
the key 'in-memory' for the current session.